Promote eagerlyResolveLoopFns to the workflow package
It walks WorkflowDefinitions resolving loop while/carry refs against a
LoopFnRegistry -- a pure @intx/workflow concern with no host coupling --
so it lives beside enumerateInlineLoopBodies rather than private to the
workflow-host run child.
