+++ packages/workflow/src/ontrigger-bodies.ts
@@ -15,6 +15,7 @@
 // layers on elsewhere.
 
 import type { Primitive, WorkflowDefinition } from "./definition/index";
+import type { LoopFnRegistry } from "./runtime/env";
 
 export interface ExtractedOnTriggerBody {
   /** The body's ref -- `<workflowId>__<stepId>` -- and the id of `definition`. */
@@ -118,6 +119,32 @@ export function enumerateInlineLoopBodies(
     bodies.push(...enumerateInlineLoopBodies(definition));
   }
   return bodies;
+}
+
+/**
+ * Force-resolve every loop `while`/`carry` ref reachable from these definitions
+ * against the registry, so a missing loop fn surfaces at establish rather than
+ * when the loop is first driven mid-run. Recurses into a loop's inline body (a
+ * nested loop resolves against the same shared registry). The caller passes the
+ * lifted onTrigger/childWorkflow bodies separately, since those are `{ ref }` in
+ * the enclosing definition and this walk does not descend into them.
+ */
+export function eagerlyResolveLoopFns(
+  definitions: readonly WorkflowDefinition[],
+  loopFns: LoopFnRegistry,
+): void {
+  const visit = (def: WorkflowDefinition): void => {
+    for (const step of Object.values(def.steps)) {
+      if (step.kind === "loop") {
+        // Each call throws (fail closed) if the ref names no export, or an
+        // export that is not a function.
+        loopFns(step.while);
+        loopFns(step.carry);
+        visit(step.body);
+      }
+    }
+  };
+  for (const def of definitions) visit(def);
 }
 
 export interface ExtractedChildWorkflowBody {
